requirements.txt
arrow==0.12.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
pyserial==3.4
python-dateutil==2.7.0
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
